feat: balance tracker by partition facet
