+++ packages/ats/contracts/contracts/facets/layer_1/ERC1400/ERC1410/ERC1410Read.sol
@@ -6,32 +6,10 @@ import { ERC1410StorageWrapper } from "../../../../domain/asset/ERC1410StorageWr
 import { ERC1594StorageWrapper } from "../../../../domain/asset/ERC1594StorageWrapper.sol";
 import { Modifiers } from "../../../../services/Modifiers.sol";
 import { PauseStorageWrapper } from "../../../../domain/core/PauseStorageWrapper.sol";
-import { TimeTravelStorageWrapper } from "../../../../test/testTimeTravel/timeTravel/TimeTravelStorageWrapper.sol";
 import { IPause } from "../../../../facets/layer_1/pause/IPause.sol";
 import { Eip1066 } from "../../../../constants/eip1066.sol";
 
 abstract contract ERC1410Read is IERC1410Read, Modifiers {
-    function balanceOfAt(address _tokenHolder, uint256 _timestamp) external view returns (uint256) {
-        return ERC1410StorageWrapper.balanceOfAdjustedAt(_tokenHolder, _timestamp);
-    }
-
-    function balanceOfByPartition(bytes32 _partition, address _tokenHolder) external view returns (uint256) {
-        return
-            ERC1410StorageWrapper.balanceOfByPartitionAdjustedAt(
-                _partition,
-                _tokenHolder,
-                TimeTravelStorageWrapper.getBlockTimestamp()
-            );
-    }
-
-    function totalSupplyByPartition(bytes32 _partition) external view returns (uint256) {
-        return
-            ERC1410StorageWrapper.totalSupplyByPartitionAdjustedAt(
-                _partition,
-                TimeTravelStorageWrapper.getBlockTimestamp()
-            );
-    }
-
     function partitionsOf(address _tokenHolder) external view returns (bytes32[] memory) {
         return ERC1410StorageWrapper.partitionsOf(_tokenHolder);
     }
