--- packages/ats/contracts/contracts/facets/layer_1/totalBalance/TotalBalanceFacet.sol
@@ -1,25 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-pragma solidity >=0.8.0 <0.9.0;
-
-import { ITotalBalance } from "./ITotalBalance.sol";
-import { TotalBalance } from "./TotalBalance.sol";
-import { IStaticFunctionSelectors } from "../../../infrastructure/proxy/IStaticFunctionSelectors.sol";
-import { _TOTAL_BALANCE_RESOLVER_KEY } from "../../../constants/resolverKeys.sol";
-
-contract TotalBalanceFacet is TotalBalance, IStaticFunctionSelectors {
-    function getStaticResolverKey() external pure override returns (bytes32 staticResolverKey_) {
-        staticResolverKey_ = _TOTAL_BALANCE_RESOLVER_KEY;
-    }
-
-    function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
-        uint256 selectorIndex;
-        staticFunctionSelectors_ = new bytes4[](1);
-        staticFunctionSelectors_[selectorIndex++] = this.getTotalBalanceForByPartition.selector;
-    }
-
-    function getStaticInterfaceIds() external pure virtual override returns (bytes4[] memory staticInterfaceIds_) {
-        staticInterfaceIds_ = new bytes4[](1);
-        uint256 selectorsIndex;
-        staticInterfaceIds_[selectorsIndex++] = type(ITotalBalance).interfaceId;
-    }
-}
