+++ packages/ats/contracts/scripts/domain/orchestratorLibraries.ts
@@ -126,10 +126,10 @@ export const LIBRARY_DEPENDENT_FACETS: Record<string, Array<keyof typeof LIBRARY
   ClearingHoldCreationFacet: ["clearingOps", "clearingProtectedOps"],
   // ClearingReadOps dependencies - clearing read operations
   ClearingReadFacet: ["clearingReadOps"],
-  // SnapshotsFacet + TotalBalanceFacet + BalanceTrackerFacet depend on SnapshotsStorageWrapper which uses ClearingReadOps
+  // SnapshotsFacet + BalanceTrackerFacet + BalanceTrackerByPartitionFacet depend on SnapshotsStorageWrapper which uses ClearingReadOps
   SnapshotsFacet: ["clearingReadOps"],
-  TotalBalanceFacet: ["clearingReadOps"],
   BalanceTrackerFacet: ["clearingReadOps"],
+  BalanceTrackerByPartitionFacet: ["clearingReadOps"],
   // Layer 2/3 Bond read facets transitively reach ClearingReadOps via SnapshotsStorageWrapper
   BondUSAReadFacet: ["clearingReadOps"],
   BondUSAReadFixedRateFacet: ["clearingReadOps"],
