--- packages/ats/contracts/test/contracts/integration/layer_1/totalBalance/totalBalance.test.ts
@@ -1,269 +1,0 @@
-// SPDX-License-Identifier: Apache-2.0
-
-import { expect } from "chai";
-import { ethers } from "hardhat";
-import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers.js";
-import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
-import { type ResolverProxy, type IAsset } from "@contract-types";
-import { ATS_ROLES, ADDRESS_ZERO, EMPTY_HEX_BYTES, EMPTY_STRING, ZERO } from "@scripts";
-import { deployEquityTokenFixture, executeRbac, MAX_UINT256 } from "@test";
-
-const _DEFAULT_PARTITION = "0x0000000000000000000000000000000000000000000000000000000000000001";
-const _SECOND_PARTITION = "0x0000000000000000000000000000000000000000000000000000000000000002";
-const _AMOUNT = 1000;
-const EMPTY_VC_ID = EMPTY_STRING;
-
-interface ClearingOperation {
-  partition: string;
-  expirationTimestamp: number;
-  data: string;
-}
-
-interface Hold {
-  amount: bigint | number;
-  expirationTimestamp: bigint | number;
-  escrow: string;
-  to: string;
-  data: string;
-}
-
-describe("Total Balance Tests", () => {
-  let diamond: ResolverProxy;
-  let signer_A: HardhatEthersSigner;
-  let signer_B: HardhatEthersSigner;
-  let signer_C: HardhatEthersSigner;
-  let signer_D: HardhatEthersSigner;
-
-  let asset: IAsset;
-
-  const ONE_YEAR_IN_SECONDS = 365 * 24 * 60 * 60;
-  let currentTimestamp = 0;
-  let expirationTimestamp = 0;
-
-  async function setFacets({ diamond }: { diamond: ResolverProxy }) {
-    asset = await ethers.getContractAt("IAsset", diamond.target);
-
-    await asset.connect(signer_A).addIssuer(signer_A.address);
-    await asset.connect(signer_B).grantKyc(signer_A.address, EMPTY_VC_ID, ZERO, MAX_UINT256, signer_A.address);
-    await asset.connect(signer_B).grantKyc(signer_B.address, EMPTY_VC_ID, ZERO, MAX_UINT256, signer_A.address);
-    await asset.connect(signer_B).grantKyc(signer_C.address, EMPTY_VC_ID, ZERO, MAX_UINT256, signer_A.address);
-  }
-
-  async function deployEquity(multiPartition: boolean) {
-    const base = await deployEquityTokenFixture({
-      equityDataParams: {
-        securityData: {
-          isMultiPartition: multiPartition,
-          clearingActive: false,
-        },
-      },
-    });
-
-    diamond = base.diamond;
-    signer_A = base.deployer;
-    signer_B = base.user1;
-    signer_C = base.user2;
-    signer_D = base.user3;
-
-    asset = await ethers.getContractAt("IAsset", diamond.target);
-    await executeRbac(asset, [
-      {
-        role: ATS_ROLES.ISSUER_ROLE,
-        members: [signer_B.address],
-      },
-      {
-        role: ATS_ROLES.LOCKER_ROLE,
-        members: [signer_C.address],
-      },
-      {
-        role: ATS_ROLES.FREEZE_MANAGER_ROLE,
-        members: [signer_D.address],
-      },
-      {
-        role: ATS_ROLES.KYC_ROLE,
-        members: [signer_B.address],
-      },
-      {
-        role: ATS_ROLES.SSI_MANAGER_ROLE,
-        members: [signer_A.address],
-      },
-      {
-        role: ATS_ROLES.CLEARING_ROLE,
-        members: [signer_A.address],
-      },
-      {
-        role: ATS_ROLES.CLEARING_VALIDATOR_ROLE,
-        members: [signer_A.address],
-      },
-    ]);
-
-    await setFacets({ diamond });
-  }
-
-  beforeEach(async () => {
-    currentTimestamp = (await ethers.provider.getBlock("latest"))!.timestamp;
-    expirationTimestamp = currentTimestamp + ONE_YEAR_IN_SECONDS;
-  });
-
-  describe("Multi-partition enabled", () => {
-    beforeEach(async () => {
-      await loadFixture(deployEquity.bind(null, true));
-    });
-
-    it("GIVEN multi-partition equity with locked, held, and cleared tokens WHEN getTotalBalanceFor and getTotalBalanceForByPartition THEN returns correct total balance", async () => {
-      const tokenHolder = signer_A.address;
-
-      // Amounts for DEFAULT partition
-      const defaultMintAmount = 600;
-      const defaultLockAmount = 100;
-      const defaultHoldAmount = 150;
-      const defaultClearAmount = 50;
-
-      // Amounts for SECOND partition
-      const secondMintAmount = 400;
-      const secondLockAmount = 80;
-      const secondHoldAmount = 70;
-      const secondClearAmount = 30;
-
-      // Mint tokens in DEFAULT partition
-      await asset.connect(signer_B).issueByPartition({
-        partition: _DEFAULT_PARTITION,
-        tokenHolder: tokenHolder,
-        value: defaultMintAmount,
-        data: EMPTY_HEX_BYTES,
-      });
-
-      // Mint tokens in SECOND partition
-      await asset.connect(signer_B).issueByPartition({
-        partition: _SECOND_PARTITION,
-        tokenHolder: tokenHolder,
-        value: secondMintAmount,
-        data: EMPTY_HEX_BYTES,
-      });
-
-      // Lock tokens in DEFAULT partition
-      await asset
-        .connect(signer_C)
-        .lockByPartition(_DEFAULT_PARTITION, defaultLockAmount, tokenHolder, expirationTimestamp);
-
-      // Lock tokens in SECOND partition
-      await asset
-        .connect(signer_C)
-        .lockByPartition(_SECOND_PARTITION, secondLockAmount, tokenHolder, expirationTimestamp);
-
-      // Hold tokens in DEFAULT partition
-      const holdDefault: Hold = {
-        amount: BigInt(defaultHoldAmount),
-        expirationTimestamp: BigInt(expirationTimestamp),
-        escrow: signer_B.address,
-        to: ADDRESS_ZERO,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset.connect(signer_A).createHoldByPartition(_DEFAULT_PARTITION, holdDefault);
-
-      // Hold tokens in SECOND partition
-      const holdSecond: Hold = {
-        amount: BigInt(secondHoldAmount),
-        expirationTimestamp: BigInt(expirationTimestamp),
-        escrow: signer_B.address,
-        to: ADDRESS_ZERO,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset.connect(signer_A).createHoldByPartition(_SECOND_PARTITION, holdSecond);
-
-      // Clear tokens in DEFAULT partition
-      await asset.connect(signer_A).activateClearing();
-
-      const clearingOperationDefault: ClearingOperation = {
-        partition: _DEFAULT_PARTITION,
-        expirationTimestamp: expirationTimestamp,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset
-        .connect(signer_A)
-        .clearingTransferByPartition(clearingOperationDefault, defaultClearAmount, signer_B.address);
-
-      // Clear tokens in SECOND partition
-      const clearingOperationSecond: ClearingOperation = {
-        partition: _SECOND_PARTITION,
-        expirationTimestamp: expirationTimestamp,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset
-        .connect(signer_A)
-        .clearingTransferByPartition(clearingOperationSecond, secondClearAmount, signer_B.address);
-
-      // Check getTotalBalanceFor - should return sum of all tokens across all partitions
-      const totalBalance = await asset.getTotalBalanceFor(tokenHolder);
-      const expectedTotalBalance = defaultMintAmount + secondMintAmount;
-      expect(totalBalance).to.equal(expectedTotalBalance);
-
-      // Check getTotalBalanceForByPartition for DEFAULT partition
-      const totalBalanceDefault = await asset.getTotalBalanceForByPartition(_DEFAULT_PARTITION, tokenHolder);
-      expect(totalBalanceDefault).to.equal(defaultMintAmount);
-
-      // Check getTotalBalanceForByPartition for SECOND partition
-      const totalBalanceSecond = await asset.getTotalBalanceForByPartition(_SECOND_PARTITION, tokenHolder);
-      expect(totalBalanceSecond).to.equal(secondMintAmount);
-    });
-  });
-
-  describe("Single partition (no multi-partition)", () => {
-    beforeEach(async () => {
-      await loadFixture(deployEquity.bind(null, false));
-    });
-
-    it("GIVEN single partition equity with locked, held, cleared, and frozen tokens WHEN getTotalBalanceFor and getTotalBalanceForByPartition THEN returns correct total balance", async () => {
-      const tokenHolder = signer_A.address;
-      const totalMintAmount = 1000;
-
-      // Amounts for operations
-      const lockAmount = 100;
-      const holdAmount = 150;
-      const clearAmount = 50;
-      const freezeAmount = 200;
-
-      // Mint tokens in DEFAULT partition
-      await asset.connect(signer_B).issueByPartition({
-        partition: _DEFAULT_PARTITION,
-        tokenHolder: tokenHolder,
-        value: totalMintAmount,
-        data: EMPTY_HEX_BYTES,
-      });
-
-      // Lock tokens
-      await asset.connect(signer_C).lockByPartition(_DEFAULT_PARTITION, lockAmount, tokenHolder, expirationTimestamp);
-
-      // Hold tokens
-      const hold: Hold = {
-        amount: BigInt(holdAmount),
-        expirationTimestamp: BigInt(expirationTimestamp),
-        escrow: signer_B.address,
-        to: ADDRESS_ZERO,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset.connect(signer_A).createHoldByPartition(_DEFAULT_PARTITION, hold);
-
-      // Clear tokens
-      await asset.connect(signer_A).activateClearing();
-
-      const clearingOperation: ClearingOperation = {
-        partition: _DEFAULT_PARTITION,
-        expirationTimestamp: expirationTimestamp,
-        data: EMPTY_HEX_BYTES,
-      };
-      await asset.connect(signer_A).clearingTransferByPartition(clearingOperation, clearAmount, signer_B.address);
-
-      // Freeze tokens
-      await asset.connect(signer_D).freezePartialTokens(tokenHolder, freezeAmount);
-
-      // Check getTotalBalanceFor - should return all tokens including frozen
-      const totalBalance = await asset.getTotalBalanceFor(tokenHolder);
-      expect(totalBalance).to.equal(totalMintAmount);
-
-      // Check getTotalBalanceForByPartition for DEFAULT partition
-      const totalBalanceByPartition = await asset.getTotalBalanceForByPartition(_DEFAULT_PARTITION, tokenHolder);
-      expect(totalBalanceByPartition).to.equal(totalMintAmount);
-    });
-  });
-});
